NE-156, 157, 161, 185: Updates to RelationshipLoaderAlgorithm, AttributeLoaderAlgorithm, and AddRemoveIntegrityCheckAlgorithm.  Also removed DefinitionLoaderAlgorithm, since it's now handled by the Attribute Loader.

+++ integration-test/src/test/java/com/wci/umls/server/test/rest/ProcessServiceRestNormalUseTest.java
@@ -97,6 +97,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfigJpa.setProject(project);
     processConfigJpa.setTerminology(umlsTerminology);
     processConfigJpa.setVersion(umlsVersion);
+    processConfigJpa.setType("MAINTAINENCE");
     processConfigJpa = (ProcessConfigJpa) processService
         .addProcessConfig(project.getId(), processConfigJpa, authToken);
     processConfig = processService.getProcessConfig(project.getId(),
@@ -121,6 +122,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfig1.setProject(project);
     processConfig1.setTerminology(umlsTerminology);
     processConfig1.setVersion(umlsVersion);
+    processConfig1.setType("MAINTAINENCE");
     ProcessConfigJpa addedProcessConfig1 = (ProcessConfigJpa) processService
         .addProcessConfig(project.getId(), processConfig1, authToken);
 
@@ -146,6 +148,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfig2.setProject(project);
     processConfig2.setTerminology(umlsTerminology);
     processConfig2.setVersion(umlsVersion);
+    processConfig2.setType("MAINTAINENCE");
     ProcessConfigJpa addedProcessConfig2 = (ProcessConfigJpa) processService
         .addProcessConfig(project.getId(), processConfig2, authToken);
 
@@ -374,6 +377,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfig.setTerminology(umlsTerminology);
     processConfig.setVersion(umlsVersion);
     processConfig.setTimestamp(new Date());
+    processConfig.setType("MAINTAINENCE");
     processConfig = processService.addProcessConfig(project.getId(),
         (ProcessConfigJpa) processConfig, authToken);
 
@@ -505,11 +509,12 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfig2 = new ProcessConfigJpa();
     processConfig2.setDescription("Process for testing use - long");
     processConfig2.setFeedbackEmail(null);
-    processConfig2.setName("Long Test Process");
+    processConfig2.setName("Long Test Process for Cancel and Restart");
     processConfig2.setProject(project);
     processConfig2.setTerminology(umlsTerminology);
     processConfig2.setVersion(umlsVersion);
     processConfig2.setTimestamp(new Date());
+    processConfig2.setType("MAINTAINENCE");
     processConfig2 = processService.addProcessConfig(project.getId(),
         (ProcessConfigJpa) processConfig2, authToken);
 
@@ -535,7 +540,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     algorithmConfig.setParameters(algoParameters);
 
     algorithmConfig = processService.addAlgorithmConfig(project.getId(),
-        processConfig.getId(), (AlgorithmConfigJpa) algorithmConfig, authToken);
+        processConfig2.getId(), (AlgorithmConfigJpa) algorithmConfig, authToken);
 
     processConfig2.getSteps().add(algorithmConfig);
 
@@ -556,7 +561,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
         .setParameters(new ArrayList<AlgorithmParameter>(algoParameters));
 
     algorithmConfig2 = processService.addAlgorithmConfig(project.getId(),
-        processConfig.getId(), (AlgorithmConfigJpa) algorithmConfig2, authToken);
+        processConfig2.getId(), (AlgorithmConfigJpa) algorithmConfig2, authToken);
 
     processConfig2.getSteps().add(algorithmConfig2);
 
@@ -740,6 +745,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfig.setTerminology(umlsTerminology);
     processConfig.setVersion(umlsVersion);
     processConfig.setTimestamp(new Date());
+    processConfig.setType("MAINTAINENCE");
     processConfig = processService.addProcessConfig(project.getId(),
         (ProcessConfigJpa) processConfig, authToken);
 
@@ -833,6 +839,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
     processConfig.setTerminology(umlsTerminology);
     processConfig.setVersion(umlsVersion);
     processConfig.setTimestamp(new Date());
+    processConfig.setType("MAINTAINENCE");
     processConfig = processService.addProcessConfig(project.getId(),
         (ProcessConfigJpa) processConfig, authToken);
 
@@ -949,7 +956,7 @@ public class ProcessServiceRestNormalUseTest extends ProcessServiceRestTest {
       processService.removeProcessConfig(project.getId(), processConfig.getId(),
           true, authToken);
     }
-    if (processConfig2 != null) {
+    if (processConfig2 != null && processConfig2.getId()!=null) {
       processService.removeProcessConfig(project.getId(),
           processConfig2.getId(), true, authToken);
     }
